refactor(xapiri): extract editor tab functions → tab_editor.go
Mechanical move per ADR 0014. No behaviour change.
